refactor(adapters): rewire imports + add barrels + lint enforcement
Part 2 of 2 for #26. Companion to the pure-mv commit immediately prior.

Changes (no behavioral change):

- Add per-substrate barrels: src/adapters/codex/index.ts, src/adapters/pi-dev/index.ts.
- Update src/adapters/codex/adapter.ts and pi-dev/adapter.ts imports to
  the new sibling and parent paths.
- Update src/adapters/codex/hooks/codex-hook-entry.mjs sibling import.
  Note: .mjs hook files keep their user-facing names — they are shipped
  verbatim to ~/.codex/hooks/ and referenced by that name in the
  generated install plan.
- Update src/adapters/codex/hooks/runtime.ts parent-relative imports
  (../../../ from new depth).
- Move feedback-helper.ts to shared/ (was codex-private but pi-dev
  consumes it too).
- Update src/adapters/pi-dev/path-guard.ts runtime-resolver path
  (../../policy-path-guard.ts from new depth).
- Update src/adapters/shared/index.ts and feedback-helper.ts
  parent-relative imports.
- Update test imports to deep paths (tests are allowed to reach into
  internals; only src/ enforces the barrier).
- Update src/install.ts to import configureCodexInstall from the codex
  barrel rather than directly from ./adapters/codex/config.

ESLint enforcement:

- New no-restricted-imports rule in eslint.config.js forbids src/ files
  outside adapters/{codex,pi-dev,shared}/ from deep-importing into
  those directories. Verified by adding a temp violating file and
  observing the rule flag it. Imports must go through the barrels.
- Rule is scoped to src/; tests retain freedom to reach into hook
  assets and runtime-render outputs.

Verified:

- bun run typecheck — clean
- bun test          — 208 pass, 0 fail
- bun run lint      — clean

Refs #26
