+++ src/adapters/codex/hooks/codex-hook-entry.mjs
