+++ src/adapters/codex/hooks/codex-policy-hook.mjs
