+++ test/policy-path-guard.test.ts
@@ -7,7 +7,7 @@ import { execFile } from "node:child_process";
 import { promisify } from "node:util";
 import { evaluatePathGuard, parseBashDestructivePaths } from "../src/policy-path-guard";
 import { evaluateSomaPolicy } from "../src/policy";
-import { renderPathGuardExtension } from "../src/adapters/pi-dev-path-guard";
+import { renderPathGuardExtension } from "../src/adapters/pi-dev/path-guard";
 import { bootstrapSomaHome } from "../src/soma-home";
 import { checkSomaPolicy, checkSomaPolicyBatch } from "../src/policy-audit";
